backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
grpcio==1.60.0
protobuf==4.25.1
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.1
